+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### [1.0.2] - 2026-05-15
 
+- feat: allow comments in relay_acl_allow
 - deps(all): bump versions
 - test: runner is now node:test
 - test: fix for testing registered hooks